[AutoOnboarding] OnFabricRemoved event is not handled
[Problem]
OnFabricRemoved event is not handled on AutoOnboarding scenario.
The event occurs when any device is removed from SmartThings.

[Cause]
exitMsgBox_enable variable that the boolean value is used when ioter blocks the process removing device itself during repeat-onboarding test

[Measure]
Remove a code setting the variable as True to show up the exit dialog